Rendering form as separate component
