+++ frontend/src/modules/Metadata_Forms/components/MetadataFormPreview.js
@@ -1,18 +1,14 @@
 import { useDispatch } from 'react-redux';
 
 import PropTypes from 'prop-types';
-import { Box, Card, CardContent, Grid, Typography } from '@mui/material';
+import { Box } from '@mui/material';
 import React, { useEffect, useState } from 'react';
 import { SET_ERROR } from '../../../globalErrors';
 import { useClient } from '../../../services';
 import { getMetadataForm } from '../services';
 import CircularProgress from '@mui/material/CircularProgress';
-import {
-  FreeInputField,
-  BooleanField,
-  DropDownField,
-  GlossaryTermField
-} from './fields';
+
+import { RenderedMetadataForm } from './renderedMetadataForm';
 
 export const MetadataFormPreview = (props) => {
   const client = useClient();
@@ -45,32 +41,6 @@ export const MetadataFormPreview = (props) => {
       );
     }
   }, [client, dispatch]);
-  const getFieldElement = (field) => {
-    if (field.type === 'Boolean') {
-      return <BooleanField field={field} />;
-    }
-
-    if (field.type === 'Glossary Term') {
-      return (
-        <GlossaryTermField field={field} client={client} dispatch={dispatch} />
-      );
-    }
-
-    if (
-      ['Integer', 'String'].includes(field.type) &&
-      (!field.possibleValues || field.possibleValues.length === 0)
-    ) {
-      return <FreeInputField field={field} />;
-    }
-
-    if (
-      ['Integer', 'String'].includes(field.type) &&
-      field.possibleValues &&
-      field.possibleValues.length > 0
-    ) {
-      return <DropDownField field={field} />;
-    }
-  };
 
   if (loading) {
     return (
@@ -88,33 +58,7 @@ export const MetadataFormPreview = (props) => {
     );
   }
 
-  return (
-    <Card>
-      {fields.map((field) => (
-        <CardContent>
-          <Grid container spacing={2}>
-            <Grid item lg={3} xl={3} xs={6}>
-              {getFieldElement(field)}
-            </Grid>
-            <Grid
-              item
-              lg={9}
-              xl={9}
-              xs={18}
-              sx={{ display: 'flex', alignItems: 'center' }}
-            >
-              <Typography variant="subtitle2" color="textPrimary">
-                {field.required && (
-                  <span style={{ color: 'red' }}>{'Required. '}</span>
-                )}
-                <span>{field.description}</span>
-              </Typography>
-            </Grid>
-          </Grid>
-        </CardContent>
-      ))}
-    </Card>
-  );
+  return <RenderedMetadataForm fields={fields} />;
 };
 
 MetadataFormPreview.propTypes = {
