+++ frontend/src/modules/Metadata_Forms/components/index.js
@@ -5,3 +5,4 @@ export * from './MetadataFormPreview';
 export * from './MetadataFormEnforcement';
 export * from './MetadataFormFields';
 export * from './fields';
+export * from './renderedMetadataForm';
